requirements.txt
fastapi[standard]
psycopg2-binary
pwdlib
SQLAlchemy
pydantic-settings
pyjwt